+++ dst/app/routes/payment/paymentCard.js
@@ -66,14 +66,33 @@ paymentCardPaymentRouter.post('/check', permitScopes_1.default(['transactions'])
  * 口座確保
  */
 // tslint:disable-next-line:use-default-type-parameter
-paymentCardPaymentRouter.post('/authorize', permitScopes_1.default(['transactions']), rateLimit_1.default, ...[
-    express_validator_1.body('object')
+paymentCardPaymentRouter.post('/authorize', permitScopes_1.default(['transactions']), rateLimit_1.default, 
+// 互換性維持のため
+(req, _, next) => {
+    var _a, _b;
+    if (req.body.object === undefined || req.body.object === null) {
+        req.body.object = {};
+    }
+    if (typeof req.body.object.notes === 'string') {
+        req.body.object.description = req.body.object.notes;
+    }
+    if (typeof ((_b = (_a = req.body.object) === null || _a === void 0 ? void 0 : _a.fromAccount) === null || _b === void 0 ? void 0 : _b.accountNumber) === 'string') {
+        if (req.body.object.fromLocation === undefined || req.body.object.fromLocation === null) {
+            req.body.object.fromLocation = {};
+        }
+        req.body.object.fromLocation.identifier = req.body.object.fromAccount.accountNumber;
+    }
+    next();
+}, ...[
+    express_validator_1.body('object.paymentMethod')
         .not()
-        .isEmpty(),
+        .isEmpty()
+        .withMessage(() => 'required')
+        .isString(),
     express_validator_1.body('object.amount')
         .not()
         .isEmpty()
-        .withMessage((_, __) => 'required')
+        .withMessage(() => 'required')
         .isInt(),
     express_validator_1.body('object.additionalProperty')
         .optional()
@@ -103,112 +122,88 @@ paymentCardPaymentRouter.post('/authorize', permitScopes_1.default(['transaction
 }), 
 // tslint:disable-next-line:max-func-body-length
 (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
-    var _a;
+    var _a, _b, _c, _d, _e, _f;
     try {
         const actionRepo = new cinerino.repository.Action(mongoose.connection);
         const transactionRepo = new cinerino.repository.Transaction(mongoose.connection);
-        let fromLocation = req.body.object.fromLocation;
+        let paymentCard;
+        const paymentMethodType = (_a = req.body.object) === null || _a === void 0 ? void 0 : _a.paymentMethod;
         // トークン化された口座情報でリクエストされた場合、実口座情報へ変換する
-        if (typeof fromLocation === 'string') {
+        if (typeof ((_b = req.body.object) === null || _b === void 0 ? void 0 : _b.fromLocation) === 'string') {
             const accountOwnershipInfo = yield cinerino.service.code.verifyToken({
                 project: req.project,
                 agent: req.agent,
-                token: fromLocation,
+                token: req.body.object.fromLocation,
                 secret: process.env.TOKEN_SECRET,
                 issuer: process.env.RESOURCE_SERVER_IDENTIFIER
             })({ action: new cinerino.repository.Action(mongoose.connection) });
-            const paymentCard = accountOwnershipInfo.typeOfGood;
-            fromLocation = {
-                typeOf: paymentCard.typeOf,
-                identifier: paymentCard.identifier
+            const typeOfGood = accountOwnershipInfo.typeOfGood;
+            paymentCard = {
+                typeOf: typeOfGood.typeOf,
+                identifier: typeOfGood.identifier
             };
         }
         else {
-            const accessCode = fromLocation === null || fromLocation === void 0 ? void 0 : fromLocation.accessCode;
-            if (typeof accessCode === 'string') {
-                // アクセスコード情報があれば、認証
-                const serviceOutputService = new cinerino.chevre.service.ServiceOutput({
-                    endpoint: cinerino.credentials.chevre.endpoint,
-                    auth: chevreAuthClient
-                });
-                const searchPaymentCardResult = yield serviceOutputService.search({
-                    limit: 1,
-                    page: 1,
-                    project: { typeOf: req.project.typeOf, id: req.project.id },
-                    typeOf: { $eq: fromLocation === null || fromLocation === void 0 ? void 0 : fromLocation.typeOf },
-                    identifier: { $eq: fromLocation === null || fromLocation === void 0 ? void 0 : fromLocation.identifier },
-                    accessCode: { $eq: accessCode }
-                });
-                if (searchPaymentCardResult.data.length === 0) {
-                    throw new cinerino.factory.errors.NotFound('PaymentCard');
+            const accessCode = (_d = (_c = req.body.object) === null || _c === void 0 ? void 0 : _c.fromLocation) === null || _d === void 0 ? void 0 : _d.accessCode;
+            const accountIdentifier = (_f = (_e = req.body.object) === null || _e === void 0 ? void 0 : _e.fromLocation) === null || _f === void 0 ? void 0 : _f.identifier;
+            if (typeof accountIdentifier === 'string') {
+                if (typeof accessCode === 'string') {
+                    // アクセスコード情報があれば、認証
+                    const serviceOutputService = new cinerino.chevre.service.ServiceOutput({
+                        endpoint: cinerino.credentials.chevre.endpoint,
+                        auth: chevreAuthClient
+                    });
+                    const searchPaymentCardResult = yield serviceOutputService.search({
+                        limit: 1,
+                        page: 1,
+                        project: { typeOf: req.project.typeOf, id: req.project.id },
+                        typeOf: { $eq: paymentMethodType },
+                        identifier: { $eq: accountIdentifier },
+                        accessCode: { $eq: accessCode }
+                    });
+                    if (searchPaymentCardResult.data.length === 0) {
+                        throw new cinerino.factory.errors.NotFound('PaymentCard');
+                    }
+                    const paymetCard = searchPaymentCardResult.data.shift();
+                    paymentCard = {
+                        typeOf: paymetCard.typeOf,
+                        identifier: paymetCard.identifier
+                    };
                 }
-                const paymetCard = searchPaymentCardResult.data.shift();
-                fromLocation = {
-                    typeOf: paymetCard.typeOf,
-                    identifier: paymetCard.identifier
-                };
-            }
-            else {
-                fromLocation = undefined;
-                // アクセスコード情報なし、かつ、会員の場合、所有権を確認
-                // 口座に所有権があるかどうか確認
-                // const ownershipInfoRepo = new cinerino.repository.OwnershipInfo(mongoose.connection);
-                // const count = await ownershipInfoRepo.count<cinerino.factory.ownershipInfo.AccountGoodType.Account>({
-                //     limit: 1,
-                //     ownedBy: { id: req.user.sub },
-                //     ownedFrom: new Date(),
-                //     ownedThrough: new Date(),
-                //     typeOfGood: {
-                //         typeOf: cinerino.factory.ownershipInfo.AccountGoodType.Account,
-                //         accountType: fromLocation.accountType,
-                //         accountNumber: fromLocation.accountNumber
-                //     }
-                // });
-                // if (count === 0) {
-                //     throw new cinerino.factory.errors.Forbidden('From Account access forbidden');
-                // }
+                else {
+                    // アクセスコード情報なし、かつ、会員の場合、所有権を確認
+                    if (typeof req.user.username === 'string') {
+                        // 口座に所有権があるかどうか確認
+                        const ownershipInfoRepo = new cinerino.repository.OwnershipInfo(mongoose.connection);
+                        const count = yield ownershipInfoRepo.count({
+                            limit: 1,
+                            ownedBy: { id: req.user.sub },
+                            ownedFrom: new Date(),
+                            ownedThrough: new Date(),
+                            typeOfGood: {
+                                typeOf: paymentMethodType,
+                                accountNumber: { $eq: accountIdentifier }
+                            }
+                        });
+                        if (count === 0) {
+                            throw new cinerino.factory.errors.Forbidden('From Account access forbidden');
+                        }
+                        paymentCard = { typeOf: paymentMethodType, identifier: accountIdentifier };
+                    }
+                }
             }
         }
-        if (fromLocation === undefined) {
-            // 入金処理は禁止
+        if (paymentCard === undefined) {
             throw new cinerino.factory.errors.ArgumentNull('From Location');
         }
-        // 注文取引、かつ、toAccount未指定の場合、販売者の口座を検索して、toAccountにセット
-        // if (toLocation === undefined) {
-        //     const transaction = await transactionRepo.findById({
-        //         typeOf: req.body.purpose.typeOf,
-        //         id: <string>req.body.purpose.id
-        //     });
-        //     if (transaction.typeOf === cinerino.factory.transactionType.PlaceOrder) {
-        //         const seller = await sellerRepo.findById({
-        //             id: transaction.seller.id
-        //         });
-        //         if (seller.paymentAccepted === undefined) {
-        //             throw new cinerino.factory.errors.Argument('object', 'Account payment not accepted');
-        //         }
-        //         const accountPaymentsAccepted =
-        //             <cinerino.factory.seller.IPaymentAccepted<cinerino.factory.paymentMethodType.PaymentCard>[]>
-        //             seller.paymentAccepted.filter((a) => a.paymentMethodType === cinerino.factory.paymentMethodType.PaymentCard);
-        //         const paymentAccepted = accountPaymentsAccepted.find((a) => a.accountType === accountType);
-        //         // tslint:disable-next-line:no-single-line-block-comment
-        //         /* istanbul ignore if */
-        //         if (paymentAccepted === undefined) {
-        //             throw new cinerino.factory.errors.Argument('object', `${accountType} payment not accepted`);
-        //         }
-        //         toLocation = {
-        //             accountNumber: paymentAccepted.accountNumber,
-        //             accountType: paymentAccepted.accountType
-        //         };
-        //     }
-        // }
         const action = yield cinerino.service.payment.chevre.authorize({
             project: req.project,
             agent: { id: req.user.sub },
-            object: Object.assign(Object.assign({ typeOf: cinerino.factory.action.authorize.paymentMethod.any.ResultType.Payment, paymentMethod: (_a = req.body.object) === null || _a === void 0 ? void 0 : _a.paymentMethod, additionalProperty: (Array.isArray(req.body.object.additionalProperty))
+            object: Object.assign(Object.assign({ typeOf: cinerino.factory.action.authorize.paymentMethod.any.ResultType.Payment, paymentMethod: paymentMethodType, additionalProperty: (Array.isArray(req.body.object.additionalProperty))
                     ? req.body.object.additionalProperty.map((p) => {
                         return { name: String(p.name), value: String(p.value) };
                     })
-                    : [], amount: Number(req.body.object.amount), accountId: fromLocation.identifier }, (typeof req.body.object.name === 'string') ? { name: req.body.object.name } : undefined), (typeof req.body.object.description === 'string') ? { description: req.body.object.description } : undefined),
+                    : [], amount: Number(req.body.object.amount), accountId: paymentCard.identifier }, (typeof req.body.object.name === 'string') ? { name: req.body.object.name } : undefined), (typeof req.body.object.description === 'string') ? { description: req.body.object.description } : undefined),
             purpose: { typeOf: req.body.purpose.typeOf, id: req.body.purpose.id },
             paymentServiceType: cinerino.factory.chevre.service.paymentService.PaymentServiceType.PaymentCard
         })({