+++ dst/app/routes/people/me/ownershipInfos/accounts.js
@@ -21,7 +21,6 @@ const permitScopes_1 = require("../../../../middlewares/permitScopes");
 const rateLimit_1 = require("../../../../middlewares/rateLimit");
 const validator_1 = require("../../../../middlewares/validator");
 const redis = require("../../../../../redis");
-const USE_MONEY_TRANFER_AMOUNT_AS_NUMBER = process.env.USE_MONEY_TRANFER_AMOUNT_AS_NUMBER === '1';
 const accountsRouter = express_1.Router();
 /**
  * 口座開設
@@ -98,7 +97,7 @@ accountsRouter.post('/:accountType', permitScopes_1.default(['people.me.*']), ra
 }));
 /**
  * 口座解約
- * 口座の状態を変更するだけで、所有口座リストから削除はしない
+ * 口座の状態を変更するだけで、所有権は変更しない
  */
 accountsRouter.put('/:accountType/:accountNumber/close', permitScopes_1.default(['people.me.*']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
@@ -106,10 +105,7 @@ accountsRouter.put('/:accountType/:accountNumber/close', permitScopes_1.default(
         const projectRepo = new cinerino.repository.Project(mongoose.connection);
         yield cinerino.service.account.close({
             project: req.project,
-            typeOf: cinerino.factory.chevre.paymentMethodType.Account,
-            ownedBy: {
-                id: req.user.sub
-            },
+            ownedBy: { id: req.user.sub },
             accountNumber: req.params.accountNumber
         })({
             ownershipInfo: ownershipInfoRepo,
@@ -125,39 +121,44 @@ accountsRouter.put('/:accountType/:accountNumber/close', permitScopes_1.default(
 /**
  * 口座取引履歴検索
  */
-accountsRouter.get('/actions/moneyTransfer', permitScopes_1.default(['people.me.*']), rateLimit_1.default, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+accountsRouter.get('/actions/moneyTransfer', permitScopes_1.default(['people.me.*']), rateLimit_1.default, ...[
+    express_validator_1.query('accountNumber')
+        .not()
+        .isEmpty()
+        .withMessage(() => 'required')
+        .isString()
+], validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const ownershipInfoRepo = new cinerino.repository.OwnershipInfo(mongoose.connection);
         const projectRepo = new cinerino.repository.Project(mongoose.connection);
         let actions = yield cinerino.service.account.searchMoneyTransferActions({
             project: req.project,
-            ownedBy: {
-                id: req.user.sub
-            },
+            ownedBy: { id: req.user.sub },
             conditions: req.query,
-            typeOfGood: { typeOf: cinerino.factory.chevre.paymentMethodType.Account }
+            typeOfGood: { accountNumber: String(req.query.accountNumber) }
         })({
             ownershipInfo: ownershipInfoRepo,
             project: projectRepo
         });
         // 互換性維持対応
-        if (USE_MONEY_TRANFER_AMOUNT_AS_NUMBER) {
-            actions = actions.map((a) => {
-                var _a;
-                return Object.assign(Object.assign({}, a), { amount: (typeof a.amount === 'number') ? a.amount : Number((_a = a.amount) === null || _a === void 0 ? void 0 : _a.value) });
-            });
-        }
-        else {
-            actions = actions.map((a) => {
-                return Object.assign(Object.assign({}, a), { amount: (typeof a.amount === 'number')
-                        ? {
-                            typeOf: 'MonetaryAmount',
-                            currency: 'Point',
-                            value: a.amount
-                        }
-                        : a.amount });
-            });
-        }
+        // if (USE_MONEY_TRANFER_AMOUNT_AS_NUMBER) {
+        //     actions = actions.map((a) => {
+        //         return {
+        //             ...a,
+        //             amount: (typeof a.amount === 'number') ? a.amount : Number(a.amount?.value)
+        //         };
+        //     });
+        // } else {
+        // }
+        actions = actions.map((a) => {
+            return Object.assign(Object.assign({}, a), { amount: (typeof a.amount === 'number')
+                    ? {
+                        typeOf: 'MonetaryAmount',
+                        currency: 'Point',
+                        value: a.amount
+                    }
+                    : a.amount });
+        });
         res.json(actions);
     }
     catch (error) {