+++ src/app/routes/people.ts
@@ -13,6 +13,14 @@ import permitScopes from '../middlewares/permitScopes';
 import rateLimit from '../middlewares/rateLimit';
 import validator from '../middlewares/validator';
 
+const chevreAuthClient = new cinerino.chevre.auth.ClientCredentials({
+    domain: <string>process.env.CHEVRE_AUTHORIZE_SERVER_DOMAIN,
+    clientId: <string>process.env.CHEVRE_CLIENT_ID,
+    clientSecret: <string>process.env.CHEVRE_CLIENT_SECRET,
+    scopes: [],
+    state: ''
+});
+
 const peopleRouter = Router();
 
 /**
@@ -195,6 +203,18 @@ peopleRouter.get<ParamsDictionary>(
     validator,
     async (req, res, next) => {
         try {
+            const productService = new cinerino.chevre.service.Product({
+                endpoint: cinerino.credentials.chevre.endpoint,
+                auth: chevreAuthClient
+            });
+            const searchPaymentCardProductsResult = await productService.search({
+                limit: 100,
+                project: { id: { $eq: req.project.id } },
+                typeOf: { $eq: cinerino.factory.chevre.product.ProductType.PaymentCard }
+            });
+            const paymentCardProducts = searchPaymentCardProductsResult.data;
+            const paymentCardOutputTypes = [...new Set(paymentCardProducts.map((p) => String(p.serviceOutput?.typeOf)))];
+
             let ownershipInfos:
                 cinerino.factory.ownershipInfo.IOwnershipInfo<cinerino.factory.ownershipInfo.IGoodWithDetail>[]
                 | cinerino.factory.ownershipInfo.IOwnershipInfo<cinerino.factory.ownershipInfo.IGood>[];
@@ -211,8 +231,8 @@ peopleRouter.get<ParamsDictionary>(
             const projectRepo = new cinerino.repository.Project(mongoose.connection);
 
             const typeOfGood = <cinerino.factory.ownershipInfo.ITypeOfGoodSearchConditions>req.query.typeOfGood;
-            switch (typeOfGood.typeOf) {
-                case cinerino.factory.chevre.paymentMethodType.Account:
+            switch (true) {
+                case paymentCardOutputTypes.includes(String(typeOfGood.typeOf)):
                     ownershipInfos = await cinerino.service.account.search({
                         project: req.project,
                         conditions: searchConditions
@@ -223,7 +243,7 @@ peopleRouter.get<ParamsDictionary>(
 
                     break;
 
-                case cinerino.factory.chevre.reservationType.EventReservation:
+                case cinerino.factory.chevre.reservationType.EventReservation === typeOfGood.typeOf:
                     ownershipInfos = await cinerino.service.reservation.searchScreeningEventReservations(<any>{
                         ...searchConditions,
                         project: { typeOf: req.project.typeOf, id: req.project.id }