+++ src/app/routes/people/me/ownershipInfos/accounts.ts
@@ -5,7 +5,7 @@ import * as cinerino from '@cinerino/domain';
 import { Router } from 'express';
 // tslint:disable-next-line:no-implicit-dependencies
 import { ParamsDictionary } from 'express-serve-static-core';
-import { body } from 'express-validator';
+import { body, query } from 'express-validator';
 import { CREATED, NO_CONTENT } from 'http-status';
 import * as mongoose from 'mongoose';
 
@@ -14,8 +14,6 @@ import rateLimit from '../../../../middlewares/rateLimit';
 import validator from '../../../../middlewares/validator';
 
 import * as redis from '../../../../../redis';
-
-const USE_MONEY_TRANFER_AMOUNT_AS_NUMBER = process.env.USE_MONEY_TRANFER_AMOUNT_AS_NUMBER === '1';
 
 const accountsRouter = Router();
 
@@ -101,9 +99,10 @@ accountsRouter.post<ParamsDictionary>(
         }
     }
 );
+
 /**
  * 口座解約
- * 口座の状態を変更するだけで、所有口座リストから削除はしない
+ * 口座の状態を変更するだけで、所有権は変更しない
  */
 accountsRouter.put(
     '/:accountType/:accountNumber/close',
@@ -117,10 +116,7 @@ accountsRouter.put(
 
             await cinerino.service.account.close({
                 project: req.project,
-                typeOf: cinerino.factory.chevre.paymentMethodType.Account,
-                ownedBy: {
-                    id: req.user.sub
-                },
+                ownedBy: { id: req.user.sub },
                 accountNumber: req.params.accountNumber
             })({
                 ownershipInfo: ownershipInfoRepo,
@@ -134,6 +130,7 @@ accountsRouter.put(
         }
     }
 );
+
 /**
  * 口座取引履歴検索
  */
@@ -141,6 +138,13 @@ accountsRouter.get(
     '/actions/moneyTransfer',
     permitScopes(['people.me.*']),
     rateLimit,
+    ...[
+        query('accountNumber')
+            .not()
+            .isEmpty()
+            .withMessage(() => 'required')
+            .isString()
+    ],
     validator,
     async (req, res, next) => {
         try {
@@ -149,38 +153,36 @@ accountsRouter.get(
 
             let actions = await cinerino.service.account.searchMoneyTransferActions({
                 project: req.project,
-                ownedBy: {
-                    id: req.user.sub
-                },
+                ownedBy: { id: req.user.sub },
                 conditions: req.query,
-                typeOfGood: { typeOf: cinerino.factory.chevre.paymentMethodType.Account }
+                typeOfGood: { accountNumber: String(req.query.accountNumber) }
             })({
                 ownershipInfo: ownershipInfoRepo,
                 project: projectRepo
             });
 
             // 互換性維持対応
-            if (USE_MONEY_TRANFER_AMOUNT_AS_NUMBER) {
-                actions = actions.map((a) => {
-                    return {
-                        ...a,
-                        amount: (typeof a.amount === 'number') ? a.amount : Number(a.amount?.value)
-                    };
-                });
-            } else {
-                actions = actions.map((a) => {
-                    return {
-                        ...a,
-                        amount: (typeof a.amount === 'number')
-                            ? {
-                                typeOf: 'MonetaryAmount',
-                                currency: 'Point', // 旧データはPointしかないのでこれで十分
-                                value: a.amount
-                            }
-                            : a.amount
-                    };
-                });
-            }
+            // if (USE_MONEY_TRANFER_AMOUNT_AS_NUMBER) {
+            //     actions = actions.map((a) => {
+            //         return {
+            //             ...a,
+            //             amount: (typeof a.amount === 'number') ? a.amount : Number(a.amount?.value)
+            //         };
+            //     });
+            // } else {
+            // }
+            actions = actions.map((a) => {
+                return {
+                    ...a,
+                    amount: (typeof a.amount === 'number')
+                        ? {
+                            typeOf: 'MonetaryAmount',
+                            currency: 'Point', // 旧データはPointしかないのでこれで十分
+                            value: a.amount
+                        }
+                        : a.amount
+                };
+            });
 
             res.json(actions);
         } catch (error) {
@@ -188,4 +190,5 @@ accountsRouter.get(
         }
     }
 );
+
 export default accountsRouter;